requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
anyio=3.3.0=py38h50d1736_0
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=20.1.0=py38h96a0964_2
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
autopep8=1.5.7=pypi_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
brotlipy=0.7.0=py38h96a0964_1001
ca-certificates=2021.10.26=hecd8cb5_2
certifi=2021.10.8=py38hecd8cb5_0
cffi=1.14.6=py38h9688ba1_0
chardet=4.0.0=py38h50d1736_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
cryptography=3.4.7=py38h1fa4640_0
cycler=0.10.0=py38_0
decorator=4.4.2=py_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
et-xmlfile=1.1.0=pypi_0
freetype=2.10.4=ha233b18_0
gurobi=9.5.0=py38_0
h5py=3.6.0=py38h4a1dd59_0
hdf5=1.10.6=hdbbcd12_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.8.1=py38h50d1736_0
intel-openmp=2019.4=233
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.2=py38hecd8cb5_1
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py38hecd8cb5_0
jupyter_server=1.11.0=pyhd8ed1ab_0
jupyterlab=3.1.11=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.8.1=pyhd8ed1ab_0
kiwisolver=1.3.0=py38h23ab428_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
line_profiler=2.1.2=py38h1de35cc_0
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.2=h79c402e_3
markupsafe=2.0.1=py38h96a0964_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h181983e_0
mistune=0.8.4=py38h96a0964_1004
mkl=2019.4=233
mkl-service=2.3.0=py38h9ed2024_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
nbclassic=0.3.1=pyhd8ed1ab_1
nbclient=0.5.4=pyhd8ed1ab_0
nbconvert=6.1.0=py38h50d1736_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
ninja=1.10.2=py38hf7b0b51_0
notebook=6.4.3=pyha770c72_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=py_0
openpyxl=3.0.7=pypi_0
openssl=1.1.1l=h9ed2024_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.1.3=py38hb1e8313_0
pandoc=2.14.2=h0d85af4_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py38h5270095_0
pip=20.3=py38hecd8cb5_0
prometheus_client=0.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.6.0=pyhd3eb1b0_2
pycodestyle=2.7.0=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.3=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=py_0
pyrsistent=0.17.3=py38h96a0964_2
pysocks=1.7.1=py38h50d1736_3
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=py_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_0
pytz=2020.4=pyhd3eb1b0_0
pyzmq=20.0.0=py38h23ab428_1
readline=8.0=h1de35cc_0
requests=2.26.0=pyhd8ed1ab_0
requests-unixsocket=0.2.0=py_0
scikit-learn=0.23.2=py38h959d312_0
scipy=1.5.2=py38h2515648_0
seaborn=0.11.0=py_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=50.3.2=py38hecd8cb5_2
six=1.15.0=py38hecd8cb5_0
sniffio=1.2.0=py38h50d1736_1
sqlite=3.33.0=hffcf06c_0
terminado=0.12.1=py38h50d1736_0
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.2=pypi_0
torchaudio=0.7.2=py38
torchfunc=0.2.0=pypi_0
torchvision=0.8.2=py38_cpu
tornado=6.1=py38h9ed2024_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.6=pyhd8ed1ab_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py_1
websocket-client=0.57.0=py38h50d1736_4
wheel=0.36.0=pyhd3eb1b0_0
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.3=hb1e8313_3
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
